+++ apibuilder4elastic/custom_flow_nodes/api-builder-plugin-traffic-monitor-api-utils/test/testHandleFilterQueries.js
@@ -65,7 +65,10 @@ describe('flow-node traffic-monitor-api-utils', () => {
 
 		// Example .../search?field=uri&value=/v2/pet/findByStatus&field=method&value=GET
 		it('should succeed with multiple filter fields', async () => {
-			const { value, output } = await flowNode.handleFilterFields({ params: { field: ["uri","method", "serviceName"], value: ["/v2/pet/findByStatus","GET", "Petstore A"] }, serviceID: "instance-1", gatewayTopology: {} });
+			const { value, output } = await flowNode.handleFilterFields({ params: { 
+				field: ["uri","method", "serviceName", "vhost"], 
+				value: ["/v2/pet/findByStatus","GET", "Petstore A", "api.customer.com"] }, 
+				serviceID: "instance-1", gatewayTopology: {} });
 
 			expect(output).to.equal('next');
 			expect(value).to.be.a('object');
@@ -73,6 +76,7 @@ describe('flow-node traffic-monitor-api-utils', () => {
 				{"match": {"http.uri": { "query": "/v2/pet/findByStatus", "operator": "and" }}},
 				{"match": {"http.method": { "query": "GET" }}},
 				{"match": {"serviceContext.service": { "query": "Petstore A", "operator": "and" }}},
+				{"match": {"http.vhost.text": { "query": "api.customer.com", "operator": "and" }}},
 				{"exists": {"field": "http"}},
 				{"term": {"processInfo.serviceId": "instance-1"}}
 			  ]}});